+++ Src/utils.c
@@ -8,7 +8,7 @@
 #include "utils.h"
 #include <stdint.h>
 
-int16_t trim(int16_t in, int16_t min, int16_t max){
+float trim(float in, float min, float max){
   if(in < min){
     return min;
   }
@@ -18,6 +18,6 @@ int16_t trim(int16_t in, int16_t min, int16_t max){
   return in;
 }
 
-int16_t map(int16_t in, int16_t from_min, int16_t from_max, int16_t to_min, int16_t to_max){
+float map(float in, float from_min, float from_max, float to_min, float to_max){
   return (in - from_min) * (to_max - to_min) / (from_max - from_min) + to_min;
 }
